requirements.txt
psycopg2-binary==2.9.10
openai>=1.0.0
